feat(alessio_di_noto): fix login e register e separazione file js
