+++ src/login.html
@@ -21,6 +21,6 @@
     </form>
     <div class="message" id="loginMessage"></div>
 
-    <script src="script.js"></script>
+    <script type="module" src="login.js"></script>
   </body>
 </html>
