+++ src/register.html
@@ -6,7 +6,7 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Registrazione</title>
-    <link rel="stylesheet" href="styles.css" />
+    <link rel="stylesheet" href="style.css" />
   </head>
   <body>
     <h2>Registrazione</h2>
@@ -19,12 +19,10 @@
       <input type="email" id="email" name="email" required />
       <label for="password">Password:</label>
       <input type="password" id="password" name="password" required />
-      <label for="descrizione">Descrizione:</label>
-      <input type="text" id="descrizione" name="descrizione" />
       <button type="submit">Register</button>
     </form>
     <div class="message" id="registerMessage"></div>
 
-    <script src="script.js"></script>
+    <script type="module" src="register.js"></script>
   </body>
 </html>
